requirements.txt
altair==6.1.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
attrs==26.1.0
blinker==1.9.0
cachetools==7.1.1
certifi==2026.4.22
charset-normalizer==3.4.7
click==8.3.3
fastapi==0.136.1
gitdb==4.0.12
GitPython==3.1.50
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.14
itsdangerous==2.2.0
Jinja2==3.1.6
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
MarkupSafe==3.0.3
narwhals==2.21.0
numpy==2.4.4
ollama==0.6.2
packaging==26.2
pandas==3.0.3
pillow==12.2.0
protobuf==7.34.1
pyarrow==24.0.0
pydantic==2.13.4
pydantic_core==2.46.4
pydeck==0.9.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-multipart==0.0.28
referencing==0.37.0
requests==2.34.0
rpds-py==0.30.0
six==1.17.0
smmap==5.0.3
starlette==1.0.0
streamlit==1.57.0
tenacity==9.1.4
toml==0.10.2
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.7.0
uvicorn==0.46.0
watchdog==6.0.0
websockets==16.0
defusedxml==0.7.1
dnspython==2.8.0
greenlet==3.5.0
ipwhois==1.3.0
python-whois==0.9.6
SQLAlchemy==2.0.49
beautifulsoup4==4.14.3
brotli==1.2.0
cffi==2.0.0
cssselect2==0.9.0
duckdb==1.5.2
fonttools==4.62.1
pycparser==3.0
pydyf==0.12.1
soupsieve==2.8.3
tinycss2==1.5.1
tinyhtml5==2.1.0
weasyprint==68.1
webencodings==0.5.1
zopfli==0.4.1
Pyphen==0.17.2
